third_party/blink/web_tests/http/tests/xmlhttprequest/response-array-buffer-abort-in-loading-state-expected.txt

Test that if responseType is set to arraybuffer, XMLHttpRequest.response is null in DONE state, after abort()-ed in LOADING state.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS xhr.status is 200
PASS xhr.response is null
PASS xhr.response is null
PASS successfullyParsed is true

TEST COMPLETE
